Main-V0.5
Merge com o branch Elisa.

Cor do fundo muda com a quantidade de lixo restante.

+++ Assets/Scripts/SpawnLixo.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class SpawnLixo : MonoBehaviour
 {
     [SerializeField] GameObject[] lixos;
-    private int spawnAmount;
+    private float spawnAmount;
     private Transform tr;
     void Start(){
         tr = GetComponent<Transform>();
